+++ src/com/github/mjkuranda/retrographicsmanager/RetroGraphicsManager.java
@@ -63,11 +63,7 @@ public class RetroGraphicsManager {
     private void executeCommand(String lineArgs[]) {
         Command c = CommandFactory.get(lineArgs);
 
-        if (c == null) {
-            return;
-        }
-
-        if (!c.isValid()) {
+        if (c == null || !c.isValid()) {
             return;
         }